// src/service/model/user.ts
export interface UserInfo {
    id: string
    isHavePet: string
    email: string
    username: string
    zipcode: any
    password: string
    s3Profile: any
    firstName: any
    lastName: any
    phoneNumber: any
    isActiveEmail: boolean
    isActivePhone: boolean
    gender: any
    dob: any
    isAccessCms: boolean
    createdAt: string
    createdBy: any
    updatedAt: any
    updatedBy: any
    deletedAt: any
    deletedBy: any
    address: any
}